post module added

+++ src/app/modules/comments/comments.service.ts
@@ -3,6 +3,7 @@ import AppError from '../../errors/AppError';
 import { UserModel } from '../users/users.model';
 import { TComment } from './comments.interface';
 import { CommentModel } from './comments.model';
+import { PostModel } from '../post/post.model';
 
 // create a comment
 const createCommentIntoB = async (
@@ -11,12 +12,12 @@ const createCommentIntoB = async (
 ) => {
   const loggedInUser = await UserModel.findOne({ email: user.userEmail });
   if (!loggedInUser) {
-    throw new AppError(httpStatus.NOT_FOUND, 'This user is not found!');
+    throw new AppError(httpStatus.UNAUTHORIZED, 'Unauthorized user!');
   }
-  //   const isPostExist = await PostModel.findById(payload?.post);
-  //   if (!isPostExist) {
-  //     throw new AppError(httpStatus.NOT_FOUND, 'This post is not found!');
-  //   }
+  const isPostExist = await PostModel.findById(payload?.post);
+  if (!isPostExist) {
+    throw new AppError(httpStatus.NOT_FOUND, 'This post is not found!');
+  }
 
   payload.user = loggedInUser?._id;
   const result = await CommentModel.create(payload);
@@ -30,10 +31,18 @@ const getAllCommentsFromDB = async (postId: string) => {
 };
 
 // update comment
-const updateCommentIntoDB = async (id: string, payload: Partial<TComment>) => {
+const updateCommentIntoDB = async (
+  id: string,
+  payload: Partial<TComment>,
+  user: Record<string, unknown>,
+) => {
   const isCommentExist = await CommentModel.findById(id);
   if (!isCommentExist) {
     throw new AppError(httpStatus.NOT_FOUND, 'This comment is not found!');
+  }
+  const loggedInUser = await UserModel.findOne({ email: user.userEmail });
+  if (loggedInUser?._id.toString() !== isCommentExist?.user.toString()) {
+    throw new AppError(httpStatus.FORBIDDEN, 'Unauthorized user!');
   }
 
   const result = await CommentModel.findByIdAndUpdate(id, payload, {
@@ -43,10 +52,17 @@ const updateCommentIntoDB = async (id: string, payload: Partial<TComment>) => {
 };
 
 // delete comment
-const deleteCommentFromDB = async (id: string) => {
+const deleteCommentFromDB = async (
+  id: string,
+  user: Record<string, unknown>,
+) => {
   const isCommentExist = await CommentModel.findById(id);
   if (!isCommentExist) {
     throw new AppError(httpStatus.NOT_FOUND, 'This comment is not found!');
+  }
+  const loggedInUser = await UserModel.findOne({ email: user.userEmail });
+  if (loggedInUser?._id.toString() !== isCommentExist?.user.toString()) {
+    throw new AppError(httpStatus.FORBIDDEN, 'Unauthorized user!');
   }
 
   const result = await CommentModel.findByIdAndDelete(id);
